requirements.txt
requests
beautifulsoup4
groq
python-dotenv